Convert all prices into decimals instead of doubles

+++ JCTG.Client/Helpers/Entity/BarData.cs
@@ -4,10 +4,10 @@
     {
         public required string Timeframe { get; set; }
         public DateTime Time { get; set; }
-        public double Open { get; set; }
-        public double High { get; set; }
-        public double Low { get; set; }
-        public double Close { get; set; }
+        public decimal Open { get; set; }
+        public decimal High { get; set; }
+        public decimal Low { get; set; }
+        public decimal Close { get; set; }
         public int TickVolume { get; set; }
     }
 }
