+++ JCTG/Entity/MetatraderMessage.cs
@@ -7,9 +7,9 @@
         public int AccountID { get; set; }
         public string OrderType { get; set; }
         public string Instrument { get; set; }
-        public double Price { get; set; }
-        public double StopLoss { get; set; }
-        public double TakeProfit { get; set; }
+        public decimal Price { get; set; }
+        public decimal StopLoss { get; set; }
+        public decimal TakeProfit { get; set; }
         public long Magic { get; set; }
     }
 }
